pyproject.toml
[project]
name = "project-manager-assistant"
version = "0.1.0"
description = "An AI agent for project management tasks."
authors = [
    { name = "Your Name", email = "you@example.com" },
]
dependencies = [
    "fastapi>=0.111.1",
    "uvicorn[standard]>=0.30.3",
    "langchain>=0.2.11",
    "langchain-openai>=0.1.12",
    "langgraph>=0.1.7",
    "redis>=5.0.7",
    "rq>=1.16.2",
    "pandas>=2.2.2",
    "pydantic-settings>=2.4.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.37.0",
    "plotly>=5.23.0",
    "networkx>=3.3",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]
requires-python = ">=3.10"
readme = "README.md"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    ".", 
]
markers = [
    "integration",
    "llm_verification"
]
